requirements.txt
flask
sqlalchemy
Flask-SQLAlchemy
PyJWT